requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.0.12
cycler==0.11.0
dataclasses==0.8
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.0
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
multidict==5.2.0
numpy==1.19.5
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
patsy==0.5.2
Pillow==8.4.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
sklearn==0.0
statsmodels==0.12.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.10.1
torchmetrics==0.8.2
torchvision==0.11.2
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.6.0
